refactor (create result package)
