+++ src/main/java/study/racing/model/Name.java
@@ -1,0 +1,4 @@
+package study.racing.model;
+
+public class Name {
+}
